apps/backend/consent-intelligence/requirements.txt
fastapi
uvicorn
pydantic
pydantic-ai-slim
python-dotenv
sqlalchemy
pg8000
python-multipart
httpx
